move map controls to its own folder
