+++ applications/fishing-map/src/features/map/Map.tsx
@@ -16,11 +16,11 @@ import { LegendLayer } from '@globalfishingwatch/ui-components/dist/map-legend/M
 import { AnyGeneratorConfig } from '@globalfishingwatch/layer-composer/dist/generators/types'
 import { useClickedEventConnect, useMapTooltip, useGeneratorsConnect } from 'features/map/map.hooks'
 import { selectDataviewInstancesResolved } from 'features/workspace/workspace.selectors'
-import { selectEditing, moveCurrentRuler } from 'features/map/rulers/rulers.slice'
+import { selectEditing, moveCurrentRuler } from 'features/map/controls/rulers.slice'
+import MapInfo from 'features/map/controls/MapInfo'
+import MapControls from 'features/map/controls/MapControls'
 import { selectDebugOptions } from 'features/debug/debug.slice'
 import { ClickPopup, HoverPopup } from './Popup'
-import MapInfo from './MapInfo'
-import MapControls from './MapControls'
 import useViewport, { useMapBounds } from './map-viewport.hooks'
 import { useMapboxRef } from './map.context'
 import styles from './Map.module.css'
