+++ applications/fishing-map/src/features/map/controls/MapInfo.module.css
