+++ applications/fishing-map/src/features/map/controls/Rulers.tsx
