+++ applications/fishing-map/src/features/map/controls/rulers.slice.ts
